+++ packages/rag/src/document/html.test.ts
@@ -1,10 +1,11 @@
-import { HTMLHeaderTextSplitter, HTMLSectionSplitter } from './html';
+import { MastraDocument } from './document';
+import { HTMLHeaderTransformer, HTMLSectionTransformer } from './html';
 
 describe('HTMLHeaderTextSplitter', () => {
-  let splitter: HTMLHeaderTextSplitter;
+  let splitter: HTMLHeaderTransformer;
 
   beforeEach(() => {
-    splitter = new HTMLHeaderTextSplitter([
+    splitter = new HTMLHeaderTransformer([
       ['h1', 'Header 1'],
       ['h2', 'Header 2'],
       ['h3', 'Header 3'],
@@ -25,10 +26,10 @@ describe('HTMLHeaderTextSplitter', () => {
       </html>
     `;
 
-    const docs = await splitter.splitText(html);
+    const docs = await splitter.splitText({ text: html });
     expect(docs.length).toBeGreaterThan(1);
-    expect(docs?.[0]?.metadata['Header 1']).toBe('Main Title');
-    expect(docs?.[1]?.metadata['Header 2']).toBe('Section 1');
+    expect(docs?.[0]?.metadata()?.['Header 1']).toBe('Main Title');
+    expect(docs?.[1]?.metadata()?.['Header 2']).toBe('Section 1');
   });
 
   it('should handle nested content', async () => {
@@ -46,14 +47,14 @@ describe('HTMLHeaderTextSplitter', () => {
       </html>
     `;
 
-    const docs = await splitter.splitText(html);
-    const mainSection = docs.find(doc => doc.metadata['Header 1'] === 'Title');
-    expect(mainSection?.pageContent).toContain('Nested content');
-    expect(mainSection?.pageContent).toContain('Deeply nested content');
+    const docs = await splitter.splitText({ text: html });
+    const mainSection = docs.find(doc => doc.metadata()?.['Header 1'] === 'Title');
+    expect(mainSection?.content()).toContain('Nested content');
+    expect(mainSection?.content()).toContain('Deeply nested content');
   });
 
   it('should respect returnEachElement option', async () => {
-    const splitterWithElements = new HTMLHeaderTextSplitter([['h1', 'Header 1']], true);
+    const splitterWithElements = new HTMLHeaderTransformer([['h1', 'Header 1']], true);
 
     const html = `
       <html>
@@ -68,20 +69,20 @@ describe('HTMLHeaderTextSplitter', () => {
       </html>
     `;
 
-    const docs = await splitterWithElements.splitText(html);
+    const docs = await splitterWithElements.splitText({ text: html });
 
     expect(docs.length).toBeGreaterThan(2);
     docs.forEach(doc => {
-      expect(doc.metadata['Header 1']).toBe('Title');
+      expect(doc.metadata()?.['Header 1']).toBe('Title');
     });
   });
 });
 
 describe('HTMLSectionSplitter', () => {
-  let splitter: HTMLSectionSplitter;
+  let splitter: HTMLSectionTransformer;
 
   beforeEach(() => {
-    splitter = new HTMLSectionSplitter([
+    splitter = new HTMLSectionTransformer([
       ['h1', 'Header 1'],
       ['h2', 'Header 2'],
     ]);
@@ -103,32 +104,33 @@ describe('HTMLSectionSplitter', () => {
 
     const docs = await splitter.splitText(html);
     expect(docs.length).toBe(3);
-    expect(docs?.[0]?.metadata['Header 1']).toBe('Document Title');
-    expect(docs?.[1]?.metadata['Header 2']).toBe('First Section');
+    expect(docs?.[0]?.metadata()?.['Header 1']).toBe('Document Title');
+    expect(docs?.[1]?.metadata()?.['Header 2']).toBe('First Section');
   });
 
   it('should properly merge metadata', async () => {
     const docs = [
-      {
-        pageContent: `
+      new MastraDocument({
+        text: `
           <h1>Title 1</h1>
           <p>Content 1</p>
         `,
         metadata: { source: 'doc1' },
-      },
-      {
-        pageContent: `
+      }),
+      new MastraDocument({
+        text: `
           <h1>Title 2</h1>
           <p>Content 2</p>
         `,
         metadata: { source: 'doc2' },
-      },
+      }),
     ].map(doc => doc);
 
     const results = await splitter.splitDocuments(docs);
+
     results.forEach(doc => {
-      expect(doc.metadata).toHaveProperty('source');
-      expect(doc.metadata).toHaveProperty('Header 1');
+      expect(doc?.metadata()).toHaveProperty('source');
+      expect(doc?.metadata()).toHaveProperty('Header 1');
     });
   });
 });
