+++ packages/rag/src/document/json.ts
@@ -1,0 +1,239 @@
+import { MastraDocument } from './document';
+
+export class RecursiveJsonTransformer {
+  private maxChunkSize: number;
+  private minChunkSize: number;
+
+  constructor({ maxChunkSize = 2000, minChunkSize }: { maxChunkSize: number; minChunkSize?: number }) {
+    this.maxChunkSize = maxChunkSize;
+    this.minChunkSize = minChunkSize ?? Math.max(maxChunkSize - 200, 50);
+  }
+
+  private static jsonSize(data: Record<string, any>): number {
+    const seen = new WeakSet();
+
+    function getStringifiableData(obj: any): any {
+      if (obj === null || typeof obj !== 'object') {
+        return obj;
+      }
+
+      if (seen.has(obj)) {
+        return '[Circular]';
+      }
+
+      seen.add(obj);
+
+      if (Array.isArray(obj)) {
+        const safeArray = [];
+        for (const item of obj) {
+          safeArray.push(getStringifiableData(item));
+        }
+        return safeArray;
+      }
+
+      const safeObj: Record<string, any> = {};
+      for (const key in obj) {
+        if (Object.prototype.hasOwnProperty.call(obj, key)) {
+          safeObj[key] = getStringifiableData(obj[key]);
+        }
+      }
+      return safeObj;
+    }
+
+    const stringifiable = getStringifiableData(data);
+    return JSON.stringify(stringifiable).length;
+  }
+
+  /**
+   * Transform JSON data while handling circular references
+   */
+  public transform(data: Record<string, any>): Record<string, any> {
+    const size = RecursiveJsonTransformer.jsonSize(data);
+
+    const seen = new WeakSet();
+
+    function createSafeCopy(obj: any): any {
+      if (obj === null || typeof obj !== 'object') {
+        return obj;
+      }
+
+      if (seen.has(obj)) {
+        return '[Circular]';
+      }
+
+      seen.add(obj);
+
+      if (Array.isArray(obj)) {
+        return obj.map(item => createSafeCopy(item));
+      }
+
+      const copy: Record<string, any> = {};
+      for (const key in obj) {
+        if (Object.prototype.hasOwnProperty.call(obj, key)) {
+          copy[key] = createSafeCopy(obj[key]);
+        }
+      }
+      return copy;
+    }
+
+    return {
+      size,
+      data: createSafeCopy(data),
+    };
+  }
+
+  /**
+   * Set a value in a nested dictionary based on the given path
+   */
+  private static setNestedDict(d: Record<string, any>, path: string[], value: any): void {
+    let current = d;
+    for (const key of path.slice(0, -1)) {
+      current[key] = current[key] || {};
+      current = current[key];
+    }
+    current[path[path.length - 1]!] = value;
+  }
+
+  /**
+   * Convert lists in the JSON structure to dictionaries with index-based keys
+   */
+  private listToDictPreprocessing(data: any): any {
+    if (data && typeof data === 'object') {
+      if (Array.isArray(data)) {
+        return Object.fromEntries(data.map((item, index) => [String(index), this.listToDictPreprocessing(item)]));
+      }
+      return Object.fromEntries(Object.entries(data).map(([k, v]) => [k, this.listToDictPreprocessing(v)]));
+    }
+    return data;
+  }
+
+  /**
+   * Split json into maximum size dictionaries while preserving structure
+   */
+  private jsonSplit({
+    data,
+    currentPath = [],
+    chunks = [{}],
+  }: {
+    data: Record<string, any>;
+    currentPath?: string[];
+    chunks?: Record<string, any>[];
+  }): Record<string, any>[] {
+    if (data && typeof data === 'object' && !Array.isArray(data)) {
+      for (const [key, value] of Object.entries(data)) {
+        const newPath = [...currentPath, key];
+        const chunkSize = RecursiveJsonTransformer.jsonSize(chunks[chunks.length - 1] || {});
+        const size = RecursiveJsonTransformer.jsonSize({ [key]: value });
+        const remaining = this.maxChunkSize - chunkSize;
+
+        if (size < remaining) {
+          // Add item to current chunk
+          RecursiveJsonTransformer.setNestedDict(chunks[chunks.length - 1] || {}, newPath, value);
+        } else {
+          if (chunkSize >= this.minChunkSize) {
+            // Chunk is big enough, start a new chunk
+            chunks.push({});
+          }
+          // Iterate
+          this.jsonSplit({
+            data: typeof value === 'object' ? value : { [key]: value },
+            currentPath: newPath,
+            chunks,
+          });
+        }
+      }
+    } else {
+      RecursiveJsonTransformer.setNestedDict(chunks[chunks.length - 1] || {}, currentPath, data);
+    }
+    return chunks;
+  }
+
+  /**
+   * Splits JSON into a list of JSON chunks
+   */
+  splitJson({
+    jsonData,
+    convertLists = false,
+  }: {
+    jsonData: Record<string, any>;
+    convertLists?: boolean;
+  }): Record<string, any>[] {
+    const processedData = convertLists ? this.listToDictPreprocessing(jsonData) : jsonData;
+
+    const chunks = this.jsonSplit({ data: processedData });
+
+    if (Object.keys(chunks[chunks.length - 1] || {}).length === 0) {
+      chunks.pop();
+    }
+
+    return chunks;
+  }
+
+  private escapeNonAscii(obj: any): any {
+    if (typeof obj === 'string') {
+      return obj.replace(/[\u0080-\uffff]/g, char => {
+        return `\\u${char.charCodeAt(0).toString(16).padStart(4, '0')}`;
+      });
+    }
+
+    if (typeof obj === 'object' && obj !== null) {
+      return Object.fromEntries(Object.entries(obj).map(([key, value]) => [key, this.escapeNonAscii(value)]));
+    }
+
+    return obj;
+  }
+  /**
+   * Splits JSON into a list of JSON formatted strings
+   */
+  splitText({
+    jsonData,
+    convertLists = false,
+    ensureAscii = true,
+  }: {
+    jsonData: Record<string, any>;
+    convertLists?: boolean;
+    ensureAscii?: boolean;
+  }): string[] {
+    const chunks = this.splitJson({ jsonData, convertLists });
+
+    if (ensureAscii) {
+      const escapedChunks = chunks.map(chunk => this.escapeNonAscii(chunk));
+      return escapedChunks.map(chunk => JSON.stringify(chunk));
+    }
+
+    return chunks.map(chunk => JSON.stringify(chunk));
+  }
+
+  /**
+   * Create documents from a list of json objects
+   */
+  createDocuments({
+    texts,
+    convertLists = false,
+    ensureAscii = true,
+    metadatas,
+  }: {
+    texts: Record<string, any>[];
+    convertLists?: boolean;
+    ensureAscii?: boolean;
+    metadatas?: Record<string, any>[];
+  }): MastraDocument[] {
+    const _metadatas = metadatas || Array(texts.length).fill({});
+    const documents: MastraDocument[] = [];
+
+    texts.forEach((text, i) => {
+      const chunks = this.splitText({ jsonData: text, convertLists, ensureAscii });
+      chunks.forEach(chunk => {
+        const metadata = { ...(_metadatas[i] || {}) };
+        documents.push(
+          new MastraDocument({
+            text: chunk,
+            metadata,
+          }),
+        );
+      });
+    });
+
+    return documents;
+  }
+}
