+++ packages/rag/src/document/markdown.test.ts
@@ -1,10 +1,10 @@
-import { MarkdownTextSplitter, MarkdownHeaderTextSplitter } from './markdown';
+import { MarkdownTransformer, MarkdownHeaderTransformer } from './markdown';
 
-describe('MarkdownTextSplitter', () => {
-  let splitter: MarkdownTextSplitter;
+describe('MarkdownTransformer', () => {
+  let splitter: MarkdownTransformer;
 
   beforeEach(() => {
-    splitter = new MarkdownTextSplitter({
+    splitter = new MarkdownTransformer({
       chunkSize: 100,
       chunkOverlap: 10,
     });
@@ -24,7 +24,7 @@ describe('MarkdownTextSplitter', () => {
   - List item 1
   - List item 2`;
 
-    const chunks = splitter.splitText(text);
+    const chunks = splitter.splitText({ text });
     expect(chunks.length).toBeGreaterThan(1);
     expect(chunks[0]).toContain('# Header 1');
   });
@@ -40,16 +40,16 @@ describe('MarkdownTextSplitter', () => {
   
   Regular text after code block.`;
 
-    const chunks = splitter.splitText(text);
+    const chunks = splitter.splitText({ text });
     expect(chunks.some(chunk => chunk.includes('```javascript'))).toBe(true);
   });
 });
 
-describe('MarkdownHeaderTextSplitter', () => {
-  let splitter: MarkdownHeaderTextSplitter;
+describe('MarkdownHeaderTransformer', () => {
+  let splitter: MarkdownHeaderTransformer;
 
   beforeEach(() => {
-    splitter = new MarkdownHeaderTextSplitter([
+    splitter = new MarkdownHeaderTransformer([
       ['#', 'Header 1'],
       ['##', 'Header 2'],
       ['###', 'Header 3'],
@@ -73,14 +73,14 @@ describe('MarkdownHeaderTextSplitter', () => {
   
   Final content.`;
 
-    const docs = splitter.splitText(text);
+    const docs = splitter.splitText({ text });
 
     expect(docs.length).toBeGreaterThan(1);
-    expect(docs?.[0]?.metadata?.['Header 1']).toBe('Main Title');
+    expect(docs?.[0]?.metadata()?.['Header 1']).toBe('Main Title');
 
-    const section1 = docs.find(doc => doc.metadata['Header 2'] === 'Section 1');
+    const section1 = docs.find(doc => doc?.metadata()?.['Header 2'] === 'Section 1');
     expect(section1).toBeDefined();
-    expect(section1?.pageContent).toContain('Section 1 content');
+    expect(section1?.content()).toContain('Section 1 content');
   });
 
   it('should handle nested headers correctly', () => {
@@ -95,12 +95,12 @@ describe('MarkdownHeaderTextSplitter', () => {
   ## Section B
   Content B`;
 
-    const docs = splitter.splitText(text);
+    const docs = splitter.splitText({ text });
 
-    const subsectionDoc = docs.find(doc => doc.metadata['Header 3'] === 'Subsection A1');
+    const subsectionDoc = docs.find(doc => doc?.metadata()?.['Header 3'] === 'Subsection A1');
     expect(subsectionDoc).toBeDefined();
-    expect(subsectionDoc?.metadata['Header 1']).toBe('Top Level');
-    expect(subsectionDoc?.metadata['Header 2']).toBe('Section A');
+    expect(subsectionDoc?.metadata()?.['Header 1']).toBe('Top Level');
+    expect(subsectionDoc?.metadata()?.['Header 2']).toBe('Section A');
   });
 
   it('should handle code blocks without splitting them', () => {
@@ -113,13 +113,13 @@ describe('MarkdownHeaderTextSplitter', () => {
   
   ## Next Section`;
 
-    const docs = splitter.splitText(text);
-    const codeDoc = docs.find(doc => doc.pageContent.includes('```python'));
-    expect(codeDoc?.pageContent).toContain('print("Hello World")');
+    const docs = splitter.splitText({ text });
+    const codeDoc = docs.find(doc => doc?.content()?.includes('```python'));
+    expect(codeDoc?.content()).toContain('print("Hello World")');
   });
 
   it('should respect returnEachLine option', () => {
-    const splitterWithLines = new MarkdownHeaderTextSplitter([['#', 'Header 1']], true);
+    const splitterWithLines = new MarkdownHeaderTransformer([['#', 'Header 1']], true);
 
     const text = `# Title
   
@@ -127,21 +127,21 @@ describe('MarkdownHeaderTextSplitter', () => {
   Line 2
   Line 3`;
 
-    const docs = splitterWithLines.splitText(text);
+    const docs = splitterWithLines.splitText({ text });
     expect(docs.length).toBe(4); // Title + 3 lines
     docs.slice(1).forEach(doc => {
-      expect(doc.metadata['Header 1']).toBe('Title');
+      expect(doc.metadata()?.['Header 1']).toBe('Title');
     });
   });
 
   it('should handle stripHeaders option', () => {
-    const splitterWithHeaders = new MarkdownHeaderTextSplitter([['#', 'Header 1']], false, false);
+    const splitterWithHeaders = new MarkdownHeaderTransformer([['#', 'Header 1']], false, false);
 
     const text = `# Title
   
   Content`;
 
-    const docs = splitterWithHeaders.splitText(text);
-    expect(docs?.[0]?.pageContent).toContain('# Title');
+    const docs = splitterWithHeaders.splitText({ text });
+    expect(docs?.[0]?.content()).toContain('# Title');
   });
 });
