+++ packages/rag/src/document/types.ts
@@ -1,0 +1,28 @@
+export enum Language {
+  CPP = 'cpp',
+  GO = 'go',
+  JAVA = 'java',
+  KOTLIN = 'kotlin',
+  JS = 'js',
+  TS = 'ts',
+  PHP = 'php',
+  PROTO = 'proto',
+  PYTHON = 'python',
+  RST = 'rst',
+  RUBY = 'ruby',
+  RUST = 'rust',
+  SCALA = 'scala',
+  SWIFT = 'swift',
+  MARKDOWN = 'markdown',
+  LATEX = 'latex',
+  HTML = 'html',
+  SOL = 'sol',
+  CSHARP = 'csharp',
+  COBOL = 'cobol',
+  C = 'c',
+  LUA = 'lua',
+  PERL = 'perl',
+  HASKELL = 'haskell',
+  ELIXIR = 'elixir',
+  POWERSHELL = 'powershell',
+}
