+++ packages/rag/src/index.ts
@@ -1,4 +1,4 @@
-export * from './document';
+export * from './document/document';
 export * from './pg';
 export * from './pinecone';
 export * from './qdrant';
